Extract shared OnboardingModal for the onboarding popups
The six onboarding notification popups each duplicated the same shell:
the handleDismiss logic (markOnboardingMessageDismissed + handleCancel),
the Modal wrapper props, the Footer/ButtonContainer/Continue button, and
a copy-pasted OnboardingPrimaryButton styled-component in every .sc.js.

Introduce OnboardingModal, which owns all of that and takes the popup body
as children plus an onboardingMessageId. Each popup now only declares its
own content.
