+++ src/pages/onboarding/notifications/DailyExercisesOnboardingPopup.js
@@ -1,41 +1,20 @@
 import * as s from "./DailyExercisesOnboardingPopup.sc";
-import Modal from "../../../components/modal_shared/Modal";
 import Main from "../../../components/modal_shared/Main.sc";
-import Footer from "../../../components/modal_shared/Footer.sc";
-import ButtonContainer from "../../../components/modal_shared/ButtonContainer.sc";
-import { useContext } from "react";
-import { APIContext } from "../../../contexts/APIContext";
+import OnboardingModal from "./OnboardingModal";
 import { ONBOARDING_MESSAGE_IDS } from "../../../appConstants";
 
 export default function DailyExercisesOnboardingPopup({ open, handleCancel }) {
-  const api = useContext(APIContext);
-  const onboardingMessageId = ONBOARDING_MESSAGE_IDS.dailyExercises;
-
-  const handleDismiss = async () => {
-    if (onboardingMessageId) {
-      try {
-        await api.markOnboardingMessageDismissed(onboardingMessageId);
-      } catch (e) {
-        // ignore dismissal recording failures
-      }
-    }
-    if (handleCancel) handleCancel();
-  };
-
   return (
-    <Modal open={open} onClose={handleDismiss} wrapperBackgroundColor="var(--onboarding-modal-bg)" hideCloseButton>
+    <OnboardingModal
+      open={open}
+      handleCancel={handleCancel}
+      onboardingMessageId={ONBOARDING_MESSAGE_IDS.dailyExercises}
+    >
       <Main>
         <s.CenteredText>Your daily exercises will appear here.</s.CenteredText>
         <s.DailyExercisesImage src="/static/images/dailyExercisesOnboarding.png" alt="Daily exercises illustration" />
         <s.CenteredSecondText>P.S. They are based on your translations from the articles.</s.CenteredSecondText>
       </Main>
-      <Footer>
-        <ButtonContainer $buttonCountNum={1}>
-          <s.OnboardingPrimaryButton $onboarding onClick={handleDismiss}>
-            Continue
-          </s.OnboardingPrimaryButton>
-        </ButtonContainer>
-      </Footer>
-    </Modal>
+    </OnboardingModal>
   );
 }
