+++ src/pages/onboarding/notifications/OnboardingModal.sc.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { StyledButton } from "../../../components/allButtons.sc";
+
+const OnboardingPrimaryButton = styled(StyledButton)`
+  min-width: 190px;
+  margin: 0 auto;
+`;
+
+export { OnboardingPrimaryButton };
