+++ src/pages/onboarding/notifications/ReviewWordsPopup.sc.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { StyledButton } from "../../../components/allButtons.sc";
 
 const ReviewWordsImage = styled.img`
   width: min(100%, 300px);
@@ -20,9 +19,5 @@ const CenteredText = styled.p`
     font-weight: 500;
   }
 `;
-const OnboardingPrimaryButton = styled(StyledButton)`
-  min-width: 190px;
-  margin: 0 auto;
-`;
 
-export { ReviewWordsImage, CenteredText, OnboardingPrimaryButton };
+export { ReviewWordsImage, CenteredText };
