+++ src/pages/onboarding/notifications/TranslationOnboardingPopup.js
@@ -1,40 +1,15 @@
 import * as s from "./TranslationOnboardingPopup.sc";
-import Modal from "../../../components/modal_shared/Modal";
 import Main from "../../../components/modal_shared/Main.sc";
-import Footer from "../../../components/modal_shared/Footer.sc";
-import ButtonContainer from "../../../components/modal_shared/ButtonContainer.sc";
-import { useContext } from "react";
-import { APIContext } from "../../../contexts/APIContext";
+import OnboardingModal from "./OnboardingModal";
 import { ONBOARDING_MESSAGE_IDS } from "../../../appConstants";
 
 export default function TranslationOnboardingPopup({ open, handleCancel }) {
-  const api = useContext(APIContext);
-  const onboardingMessageId = ONBOARDING_MESSAGE_IDS.translation;
-
-  const handleDismiss = async () => {
-    if (onboardingMessageId) {
-      try {
-        await api.markOnboardingMessageDismissed(onboardingMessageId);
-      } catch (e) {
-        // ignore dismissal recording failures
-      }
-    }
-    if (handleCancel) handleCancel();
-  };
-
   return (
-    <Modal open={open} onClose={handleDismiss} wrapperBackgroundColor="var(--onboarding-modal-bg)" hideCloseButton>
+    <OnboardingModal open={open} handleCancel={handleCancel} onboardingMessageId={ONBOARDING_MESSAGE_IDS.translation}>
       <s.TranslationImage src="/static/images/translate.png" alt="Translation illustration" />
       <Main>
         <s.CenteredText>Tap on any word to translate it. Add at least 3 words to start learning them.</s.CenteredText>
       </Main>
-      <Footer>
-        <ButtonContainer $buttonCountNum={1}>
-          <s.OnboardingPrimaryButton $onboarding onClick={handleDismiss}>
-            Continue
-          </s.OnboardingPrimaryButton>
-        </ButtonContainer>
-      </Footer>
-    </Modal>
+    </OnboardingModal>
   );
 }
