+++ src/pages/onboarding/notifications/TranslationOnboardingPopup.sc.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { StyledButton } from "../../../components/allButtons.sc";
 
 const TranslationImage = styled.img`
   width: min(100%, 300px);
@@ -21,9 +20,5 @@ const CenteredText = styled.p`
     font-weight: 500;
   }
 `;
-const OnboardingPrimaryButton = styled(StyledButton)`
-  min-width: 190px;
-  margin: 0 auto;
-`;
 
-export { TranslationImage, CenteredText, OnboardingPrimaryButton };
+export { TranslationImage, CenteredText };
